Add files via upload
The dataset for this competition (both train and test) was generated from a deep learning model trained on the Kidney Stone Prediction based on Urine Analysis dataset. Feature distributions are close to, but not exactly the same, as the original. Feel free to use the original dataset as part of this competition, both to explore differences as well as to see whether incorporating the original in training improves model performance.

Files
train.csv - the training dataset; target is the likelihood of a kidney stone being present
test.csv - the test dataset; your objective is to predict the probability of target
sample_submission.csv - a sample submission file in the correct format